backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
psycopg2-binary
sqlalchemy
alembic
python-dotenv
bcrypt
passlib[bcrypt]
pyjwt
httpx
redis
aioredis
authlib
asyncpg